Strip resources.c/.h out of common.c/.h
